Add redis tests
- Separate basic test cases from backend used.
- Exercise test cases with both redis & in-memory backends.
