+++ .travis.yml
@@ -5,6 +5,9 @@ node_js:
 
 sudo: false
 
+services:
+  - redis-server
+
 notifications:
   irc:
     channels:
